+++ Gw2Sharp/WebApi/V2/Models/SkillSlot.cs
@@ -1,0 +1,78 @@
+namespace Gw2Sharp.WebApi.V2.Models
+{
+    /// <summary>
+    /// The skill slot.
+    /// </summary>
+    public enum SkillSlot
+    {
+        /// <summary>
+        /// Unknown slot.
+        /// </summary>
+        Unknown,
+
+        /// <summary>
+        /// First weapon slot.
+        /// </summary>
+        Weapon1,
+
+        /// <summary>
+        /// Second weapon slot.
+        /// </summary>
+        Weapon2,
+
+        /// <summary>
+        /// Third weapon slot.
+        /// </summary>
+        Weapon3,
+
+        /// <summary>
+        /// Fourth weapon slot.
+        /// </summary>
+        Weapon4,
+
+        /// <summary>
+        /// Fifth weapon slot.
+        /// </summary>
+        Weapon5,
+
+        /// <summary>
+        /// First profession slot.
+        /// </summary>
+        Profession1,
+
+        /// <summary>
+        /// Second profession slot.
+        /// </summary>
+        Profession2,
+
+        /// <summary>
+        /// Third profession slot.
+        /// </summary>
+        Profession3,
+
+        /// <summary>
+        /// Fourth profession slot.
+        /// </summary>
+        Profession4,
+
+        /// <summary>
+        /// Fifth profession slot.
+        /// </summary>
+        Profession5,
+
+        /// <summary>
+        /// Heal slot.
+        /// </summary>
+        Heal,
+
+        /// <summary>
+        /// Utility slot.
+        /// </summary>
+        Utility,
+
+        /// <summary>
+        /// Elite slot.
+        /// </summary>
+        Elite
+    }
+}